Update decryption.py
A couple of changes in the loop. The first thing is that I changed the names of the variables to better represent the logic flow of the decryption, were we start off with the L, followed by T, S and than the subkey subtraction. This is because it's exactly the opposite of the encryption in which we have subkey sum, S, T and L. The first z is computed outside the loop for efficiency reason. It could also works inside the loop since z is immediately rewrited aside the first iteration but it's better if it is outside. This also allow the code to be coherent with the encryption since both theoretically and practically the code is reversed. The second major change is in the parameters of the range since the start is include and the finish is excluded, in order to compute from 4 to 0 we need to change from n-1 to -1.